server/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
cryptography==41.0.4
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.4
lxml==4.9.3
MarkupSafe==2.1.3
packaging==23.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
boto3>=1.18.63,<2.0
python-dotenv>=1.0.0
redis>=5.0.1
requests>=2.31.0
